ref(onboarding): Rename wizardNew -> onboarding
