+++ src/sentry/static/sentry/app/views/onboarding/welcome.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'react-emotion';
 
 import {analytics} from 'app/utils/analytics';
-import {stepPropTypes} from 'app/views/onboarding/wizardNew';
+import {stepPropTypes} from 'app/views/onboarding/onboarding';
 import {t, tct} from 'app/locale';
 import Button from 'app/components/button';
 import SentryTypes from 'app/sentryTypes';
